Refactor project structure, update CI, and improve test reliability
This PR restructures the project to use a modern src layout, updates the Makefile and CI workflow for robust test and security checks, and ensures all imports and paths work consistently across local and CI environments.

Moves source code to src, standardizes test and asset locations
Updates Makefile and GitHub Actions to set PYTHONPATH=src and lower coverage threshold for now
Fixes Bandit config usage and artifact upload for local and CI runs
Verifies that tests, linting, formatting, and security checks all run and pass in both local and CI environments
Why:
To follow best practices for Python project structure, improve maintainability, and ensure reliable automated testing and security scanning.

How to test:

Run make test and make test-cov locally
Run the full CI pipeline (test and security jobs)
Confirm that all tests pass, coverage is reported, and security reports are generated and uploaded
