+++ src/pro_photo_processor/presets/format_optimizer.py
