+++ src/pro_photo_processor/raw/raw_processing_enhanced.py
