--- tests/__init__.py
@@ -1,1 +1,0 @@
-# Photo Post-Processing Tests Package
