+++ tests/core/test_resize_only.py
