--- tests/format_comparison/JPEG_vs_NEF_Analysis.md
@@ -1,127 +1,0 @@
-# JPEG vs NEF Processing Analysis & Automatic Optimization
-
-## Summary of Your Observation
-
-You noticed significant differences when processing the same image in both JPEG and NEF (RAW) formats:
-- **Left image**: JPEG post-processing result
-- **Right image**: NEF (RAW) post-processing result
-
-These differences are **expected and normal** due to the fundamental differences between file formats.
-
-## Why JPEG and NEF Results Differ
-
-### 1. **Dynamic Range**
-- **JPEG**: 8-bit per channel (256 levels) - limited dynamic range
-- **NEF**: 12-14 bit per channel (4,096-16,384 levels) - much wider dynamic range
-- **Impact**: NEF files can recover more detail from highlights and shadows
-
-### 2. **Color Space**
-- **JPEG**: Usually sRGB color space (smaller gamut)
-- **NEF**: Larger color gamuts (Adobe RGB or ProPhoto RGB equivalent)
-- **Impact**: NEF files can display more vibrant and accurate colors
-
-### 3. **Processing Headroom**
-- **JPEG**: Already processed by camera - limited adjustment tolerance
-- **NEF**: Raw sensor data - can handle aggressive adjustments without artifacts
-
-### 4. **Compression**
-- **JPEG**: Lossy compression - some detail already lost
-- **NEF**: Lossless compression - all original detail preserved
-
-## Automatic Optimization System (NEW!)
-
-I've implemented an **automatic format detection and optimization system** that now:
-
-### ✅ **Detects File Format Automatically**
-```
-📊 Format analysis: 2 RAW, 3 JPEG, 0 other
-🔄 Mixed formats detected - automatic optimization will choose:
-   📷 RAW files -> sports_action_raw
-   🖼️  JPEG files -> sports_action
-```
-
-### ✅ **Uses Format-Specific Presets**
-- **JPEG files**: Use moderate settings (e.g., `sports_action`)
-- **NEF files**: Use aggressive settings (e.g., `sports_action_raw`)
-
-### ✅ **Added Enhanced RAW Presets**
-New RAW-optimized presets with more aggressive processing:
-- `sports_action_raw` - Enhanced for sports NEF files
-- `portrait_dramatic_raw` - Enhanced dramatic portraits for RAW
-- `portrait_natural_raw` - Enhanced natural portraits for RAW
-- `landscape_raw` - Enhanced landscapes for RAW
-- `natural_wildlife_raw` - Enhanced wildlife for RAW
-
-## Preset Comparison: Regular vs RAW-Optimized
-
-| Parameter | sports_action (JPEG) | sports_action_raw (NEF) | Difference |
-|-----------|---------------------|------------------------|------------|
-| Exposure  | 0.1                | 0.2                    | +100% more |
-| Highlights| -15                | -20                    | +33% stronger |
-| Shadows   | 20                 | 25                     | +25% more lift |
-| Vibrance  | 18                 | 25                     | +39% more vibrant |
-| Saturation| 5                  | 10                     | +100% more |
-| Clarity   | 12                 | 18                     | +50% more detail |
-| Structure | 15                 | 22                     | +47% sharper |
-
-## What This Means for Your Workflow
-
-### 🎯 **For Best Results**
-1. **Always shoot in NEF when possible** - gives you the most processing flexibility
-2. **Use the pipeline with mixed files** - it will automatically optimize each format
-3. **Expect different results** - NEF files should look more vibrant and detailed
-
-### 🔄 **Automatic Processing**
-The pipeline now automatically:
-1. Scans your batch of files
-2. Detects JPEG vs NEF formats
-3. Applies appropriate presets for each format
-4. Shows you which optimization was chosen
-
-### ⚡ **Performance Impact**
-- **NEF files**: ~3x longer processing time but much better quality
-- **JPEG files**: Faster processing, good for quick results
-- **Mixed batches**: Optimized processing for each file type
-
-## Example Processing Output
-
-```
-📁 Found 5 image files to process
-📊 Format analysis: 2 RAW, 3 JPEG, 0 other
-🔄 Mixed formats detected - automatic optimization will choose:
-   📷 RAW files -> sports_action_raw
-   🖼️  JPEG files -> sports_action
-
-Processing 4K images with sports_action preset...
-🔄 IMG_001.NEF (RAW) -> using sports_action_raw
-   📝 IMG_001.NEF: Vibrance: 25, Structure: 22, Clarity: 18
-   📝 IMG_002.JPG: Vibrance: 18, Structure: 15, Clarity: 12
-🔄 IMG_003.NEF (RAW) -> using sports_action_raw
-   📝 IMG_003.NEF: Vibrance: 25, Structure: 22, Clarity: 18
-```
-
-## Recommendations
-
-### 🏆 **For Sports Photography**
-- **NEF files**: Always use `sports_action_raw` (or let auto-optimization choose it)
-- **JPEG files**: Use `sports_action`
-- **Mixed batches**: Just select `sports_action` - the system handles the rest
-
-### 🎨 **For Portraits**
-- **NEF files**: `portrait_dramatic_raw` for more impact, `portrait_natural_raw` for subtle
-- **JPEG files**: Standard presets work well
-
-### 🔧 **Pipeline Integration**
-The automatic optimization is now integrated into your main pipeline. Just run it normally and it will:
-1. Detect formats automatically
-2. Choose optimal presets
-3. Show you what optimizations were applied
-4. Process each file with the best settings
-
-## Next Steps
-1. **Test with your sports images**: Process both JPEG and NEF versions
-2. **Compare results**: You should see more vibrant, detailed results from NEF files
-3. **Use mixed batches**: The system will handle format optimization automatically
-4. **Fine-tune if needed**: Adjust individual presets based on your preferences
-
-Your observation about the differences between JPEG and NEF processing was spot-on, and now the system automatically optimizes for each format!
