--- tests/format_comparison/raw_vs_jpeg_guide.py
@@ -1,229 +1,0 @@
-"""
-RAW vs JPEG: The Definitive Guide for Post-Processing
-Professional recommendation for optimal photo enhancement results
-"""
-
-
-def analyze_raw_vs_jpeg_advantages():
-    """Comprehensive analysis of RAW vs JPEG for post-processing"""
-
-    print("📸 RAW vs JPEG POST-PROCESSING GUIDE")
-    print("=" * 60)
-
-    print("🏆 WINNER: RAW/NEF FILES")
-    print("Recommendation: ALWAYS use RAW files when possible")
-    print()
-
-    # Technical advantages
-    advantages = {
-        "Dynamic Range": {
-            "RAW/NEF": "12-14 bits per channel (4,096-16,384 levels)",
-            "JPEG": "8 bits per channel (256 levels)",
-            "Advantage": "RAW has 16-64x more tonal information",
-            "Impact": "Better highlight/shadow recovery, smoother gradients",
-        },
-        "Color Space": {
-            "RAW/NEF": "Full sensor gamut (wider than Adobe RGB)",
-            "JPEG": "Usually sRGB (limited color range)",
-            "Advantage": "RAW captures more colors",
-            "Impact": "More vibrant, accurate colors after processing",
-        },
-        "Processing Headroom": {
-            "RAW/NEF": "Unprocessed sensor data",
-            "JPEG": "Already processed by camera",
-            "Advantage": "RAW can handle aggressive adjustments",
-            "Impact": "No artifacts from extreme adjustments",
-        },
-        "White Balance": {
-            "RAW/NEF": "Completely adjustable after capture",
-            "JPEG": "Baked in, limited adjustment",
-            "Advantage": "RAW allows perfect color temperature",
-            "Impact": "Fix color casts, creative temperature effects",
-        },
-        "Compression": {
-            "RAW/NEF": "Lossless compression",
-            "JPEG": "Lossy compression",
-            "Advantage": "RAW preserves all original detail",
-            "Impact": "Better sharpening, detail enhancement",
-        },
-    }
-
-    for aspect, details in advantages.items():
-        print(f"🔍 {aspect.upper()}:")
-        print(f"   📷 RAW/NEF: {details['RAW/NEF']}")
-        print(f"   🖼️  JPEG:   {details['JPEG']}")
-        print(f"   ✅ Advantage: {details['Advantage']}")
-        print(f"   💡 Impact: {details['Impact']}")
-        print()
-
-
-def processing_recommendations_by_genre():
-    """Specific recommendations for different photography genres"""
-
-    print("🎯 GENRE-SPECIFIC RECOMMENDATIONS")
-    print("=" * 50)
-
-    genres = {
-        "Sports Photography": {
-            "best_format": "RAW/NEF",
-            "reasons": [
-                "Fast action needs exposure recovery",
-                "Team colors benefit from vibrance adjustments",
-                "Outdoor lighting varies dramatically",
-                "Detail enhancement for faces/equipment",
-            ],
-            "preset": "sports_action_raw",
-            "key_benefits": "Better action freeze, color pop, detail",
-        },
-        "Portrait Photography": {
-            "best_format": "RAW/NEF",
-            "reasons": [
-                "Skin tone perfection requires color headroom",
-                "Shadow/highlight balance for flattering light",
-                "Fine detail in eyes, hair, skin texture",
-                "Creative color grading possibilities",
-            ],
-            "preset": "portrait_dramatic_raw or portrait_natural_raw",
-            "key_benefits": "Perfect skin tones, dramatic lighting, fine detail",
-        },
-        "Wildlife Photography": {
-            "best_format": "RAW/NEF",
-            "reasons": [
-                "Fur/feather texture needs structure enhancement",
-                "Challenging lighting conditions",
-                "Natural color accuracy critical",
-                "Often need exposure compensation",
-            ],
-            "preset": "natural_wildlife_raw",
-            "key_benefits": "Texture detail, natural colors, exposure flexibility",
-        },
-        "Landscape Photography": {
-            "best_format": "RAW/NEF",
-            "reasons": [
-                "Sky/foreground exposure differences",
-                "Color grading for mood",
-                "Fine detail in textures",
-                "HDR-like processing possible",
-            ],
-            "preset": "landscape_raw",
-            "key_benefits": "Sky recovery, vibrant colors, fine detail",
-        },
-        "Street Photography": {
-            "best_format": "RAW/NEF",
-            "reasons": [
-                "Mixed lighting conditions",
-                "Quick exposure compensation needed",
-                "Color grading for mood",
-                "Shadow detail in urban environments",
-            ],
-            "preset": "portrait_natural_raw or sports_action_raw",
-            "key_benefits": "Lighting flexibility, mood enhancement",
-        },
-    }
-
-    for genre, details in genres.items():
-        print(f"📷 {genre.upper()}:")
-        print(f"   🏆 Best Format: {details['best_format']}")
-        print(f"   🎨 Recommended Preset: {details['preset']}")
-        print(f"   ✨ Key Benefits: {details['key_benefits']}")
-        print("   📋 Why RAW is Better:")
-        for reason in details["reasons"]:
-            print(f"      • {reason}")
-        print()
-
-
-def when_jpeg_might_be_acceptable():
-    """Limited scenarios where JPEG might be acceptable"""
-
-    print("⚠️  WHEN JPEG MIGHT BE ACCEPTABLE")
-    print("=" * 40)
-
-    print("🔶 Limited scenarios where JPEG is OK:")
-    print("   • Social media quick posts (already small/compressed)")
-    print("   • Perfect lighting with no needed adjustments")
-    print("   • Storage space extremely limited")
-    print("   • Immediate delivery required (no time for processing)")
-    print("   • Camera doesn't support RAW")
-    print()
-
-    print("❌ JPEG is NOT recommended for:")
-    print("   • Professional work")
-    print("   • Challenging lighting")
-    print("   • Creative post-processing")
-    print("   • Printing large sizes")
-    print("   • Any important/irreplaceable photos")
-
-
-def workflow_recommendations():
-    """Optimal workflow recommendations"""
-
-    print("🔧 OPTIMAL WORKFLOW RECOMMENDATIONS")
-    print("=" * 45)
-
-    print("📋 CAMERA SETTINGS:")
-    print("   1. Shoot RAW+JPEG if storage allows")
-    print("   2. Use RAW for post-processing")
-    print("   3. Keep JPEG for quick previews/sharing")
-    print("   4. Set camera to highest quality RAW")
-    print()
-
-    print("💻 POST-PROCESSING WORKFLOW:")
-    print("   1. Import RAW files to your pipeline")
-    print("   2. Let automatic format optimization choose RAW presets")
-    print("   3. Fine-tune based on specific image needs")
-    print("   4. Export high-quality JPEG for final use")
-    print()
-
-    print("⚡ PERFORMANCE CONSIDERATIONS:")
-    print("   • RAW processing: 2-4x longer than JPEG")
-    print("   • RAW files: 3-5x larger than JPEG")
-    print("   • Better quality justifies the extra time/space")
-    print("   • Batch processing makes workflow efficient")
-
-
-def real_world_comparison():
-    """Real-world quality comparison"""
-
-    print("🌟 REAL-WORLD QUALITY COMPARISON")
-    print("=" * 40)
-
-    print("📊 Processing Capability Comparison:")
-    print()
-
-    capabilities = [
-        ("Exposure Recovery", "RAW: ±3 stops", "JPEG: ±0.5 stops"),
-        ("Highlight Recovery", "RAW: Excellent", "JPEG: Limited"),
-        ("Shadow Detail", "RAW: Excellent", "JPEG: Poor"),
-        ("Color Grading", "RAW: Full control", "JPEG: Restricted"),
-        ("White Balance", "RAW: Perfect", "JPEG: Limited"),
-        ("Noise Reduction", "RAW: Advanced", "JPEG: Basic"),
-        ("Sharpening", "RAW: Professional", "JPEG: Basic"),
-        ("Print Quality", "RAW: Excellent", "JPEG: Good"),
-    ]
-
-    print(f"{'Capability':<18} {'RAW/NEF':<20} {'JPEG':<15}")
-    print("-" * 55)
-
-    for capability, raw_quality, jpeg_quality in capabilities:
-        print(f"{capability:<18} {raw_quality:<20} {jpeg_quality:<15}")
-
-    print()
-    print("🎯 BOTTOM LINE:")
-    print("   RAW files give you 5-10x more post-processing capability")
-    print("   The quality difference is immediately visible")
-    print("   Professional photographers use RAW exclusively")
-
-
-if __name__ == "__main__":
-    analyze_raw_vs_jpeg_advantages()
-    processing_recommendations_by_genre()
-    when_jpeg_might_be_acceptable()
-    workflow_recommendations()
-    real_world_comparison()
-
-    print("\n" + "=" * 60)
-    print("🏆 FINAL RECOMMENDATION:")
-    print("   ALWAYS USE RAW/NEF FILES FOR POST-PROCESSING")
-    print("   Your pipeline automatically optimizes for RAW files")
-    print("   The quality difference is substantial and worth it")
-    print("=" * 60)
