--- tests/format_comparison/your_pipeline_guide.py
@@ -1,233 +1,0 @@
-"""
-Your Pipeline: RAW vs JPEG Processing Demonstration
-Shows how your automatic optimization system maximizes RAW file advantages
-"""
-
-import os
-import sys
-
-from utils.format_optimizer import FormatOptimizer
-from utils.photoshop_tools import PHOTOSHOP_PRESETS
-
-sys.path.append(os.path.join(os.path.dirname(__file__), "..", "..", "src"))
-
-
-def demonstrate_your_pipeline_advantages():
-    """Show how your pipeline maximizes RAW advantages"""
-
-    print("🎨 YOUR PIPELINE: RAW vs JPEG OPTIMIZATION")
-    print("=" * 60)
-
-    optimizer = FormatOptimizer()
-
-    # Show what happens with different file types
-    test_scenarios = [
-        ("Sports Action Shot", "IMG_001.NEF", "sports_action"),
-        ("Sports Action Shot", "IMG_001.JPG", "sports_action"),
-        ("Portrait", "Portrait.CR2", "portrait_dramatic"),
-        ("Portrait", "Portrait.jpeg", "portrait_dramatic"),
-        ("Wildlife", "Eagle.ARW", "natural_wildlife"),
-        ("Wildlife", "Eagle.jpg", "natural_wildlife"),
-    ]
-
-    print("🔄 AUTOMATIC PRESET OPTIMIZATION:")
-    print()
-    print(
-        f"{'Scenario':<20} {'File':<15} {'You Choose':<18} {'Pipeline Uses':<20} {'Advantage'}"
-    )
-    print("-" * 90)
-
-    for scenario, filename, requested, _ in test_scenarios:
-        optimal = optimizer.get_optimal_preset(filename, requested)
-        format_type = optimizer.detect_file_format(filename)
-
-        if format_type == "raw":
-            advantage = "🚀 RAW Power!"
-        else:
-            advantage = "📱 JPEG Quick"
-
-        print(
-            f"{scenario:<20} {filename:<15} {requested:<18} {optimal:<20} {advantage}"
-        )
-
-    print("\n" + "=" * 60)
-
-
-def show_preset_differences():
-    """Show the actual differences in processing power"""
-
-    print("⚡ PROCESSING POWER DIFFERENCES")
-    print("=" * 40)
-
-    # Compare sports action presets
-    jpeg_preset = PHOTOSHOP_PRESETS["sports_action"]
-    raw_preset = PHOTOSHOP_PRESETS["sports_action_raw"]
-
-    print("🏃‍♂️ SPORTS ACTION COMPARISON:")
-    print()
-    print(f"{'Parameter':<15} {'JPEG Preset':<12} {'RAW Preset':<12} {'RAW Advantage'}")
-    print("-" * 55)
-
-    params = [
-        "exposure",
-        "highlights",
-        "shadows",
-        "vibrance",
-        "saturation",
-        "clarity",
-        "structure",
-    ]
-
-    for param in params:
-        jpeg_val = jpeg_preset.get(param, 0)
-        raw_val = raw_preset.get(param, 0)
-
-        if jpeg_val != 0:
-            advantage_pct = ((raw_val - jpeg_val) / jpeg_val) * 100
-            advantage = f"+{advantage_pct:.0f}%"
-        else:
-            if raw_val > 0:
-                advantage = f"+{raw_val}"
-            else:
-                advantage = "-"
-
-        print(f"{param:<15} {jpeg_val:<12} {raw_val:<12} {advantage}")
-
-    print()
-    print("💡 What this means:")
-    print("   • RAW files get MUCH more aggressive processing")
-    print("   • Higher vibrance = more color pop")
-    print("   • More structure = sharper details")
-    print("   • Better highlight recovery = saved blown-out areas")
-
-
-def practical_benefits_for_your_work():
-    """Practical benefits specific to your photography"""
-
-    print("\n🎯 PRACTICAL BENEFITS FOR YOUR PHOTOGRAPHY")
-    print("=" * 50)
-
-    benefits = {
-        "Sports Photography": [
-            "Team jerseys pop with vibrant colors (+39% vibrance)",
-            "Player faces sharp even in shadows (+67% shadow lift)",
-            "Equipment details crystal clear (+47% structure)",
-            "Action freeze with proper exposure recovery",
-            "Crowd atmosphere enhanced with better colors",
-        ],
-        "Portrait Photography": [
-            "Perfect skin tones with color headroom",
-            "Eye detail enhancement without artifacts",
-            "Hair texture and fine details preserved",
-            "Creative color grading possibilities",
-            "Natural lighting balance in any condition",
-        ],
-        "Wildlife Photography": [
-            "Fur and feather texture enhancement (+80% structure)",
-            "Natural color accuracy in any lighting",
-            "Shadow detail in forest/shade conditions",
-            "No processing artifacts on fine details",
-            "Better print quality for wall art",
-        ],
-    }
-
-    for genre, benefit_list in benefits.items():
-        print(f"📷 {genre.upper()}:")
-        for benefit in benefit_list:
-            print(f"   ✅ {benefit}")
-        print()
-
-
-def your_workflow_recommendation():
-    """Specific workflow recommendation for your setup"""
-
-    print("🔧 YOUR OPTIMAL WORKFLOW")
-    print("=" * 30)
-
-    print("📋 STEP-BY-STEP PROCESS:")
-    print()
-    print("1️⃣  CAMERA SETUP:")
-    print("   • Set camera to RAW+JPEG (if storage allows)")
-    print("   • Always shoot NEF for important shots")
-    print("   • Use highest quality RAW settings")
-    print()
-
-    print("2️⃣  FILE ORGANIZATION:")
-    print("   • Transfer NEF files to your input folder")
-    print("   • Keep JPEGs for quick previews only")
-    print("   • Your pipeline handles mixed formats automatically")
-    print()
-
-    print("3️⃣  PROCESSING:")
-    print("   • Run your pipeline with any preset you want")
-    print("   • System automatically detects RAW vs JPEG")
-    print("   • RAW files get enhanced presets automatically")
-    print("   • No manual preset switching needed")
-    print()
-
-    print("4️⃣  RESULTS:")
-    print("   • NEF files: Maximum quality and detail")
-    print("   • JPEG files: Good quality, faster processing")
-    print("   • Both optimized for their capabilities")
-    print()
-
-    print("🎯 FOR BEST RESULTS:")
-    print("   📷 Sports: Always use NEF → automatic sports_action_raw")
-    print("   👤 Portraits: Always use NEF → automatic portrait_*_raw")
-    print("   🦅 Wildlife: Always use NEF → automatic natural_wildlife_raw")
-    print("   🏔️  Landscapes: Always use NEF → automatic landscape_raw")
-
-
-def storage_and_performance_reality():
-    """Reality check on storage and performance"""
-
-    print("\n📊 STORAGE & PERFORMANCE REALITY CHECK")
-    print("=" * 45)
-
-    print("💾 STORAGE REQUIREMENTS:")
-    print("   • NEF files: ~25-50MB each")
-    print("   • JPEG files: ~5-15MB each")
-    print("   • RAW takes 3-5x more space")
-    print("   • Modern storage is cheap - quality is priceless")
-    print()
-
-    print("⏱️  PROCESSING TIME:")
-    print("   • NEF processing: ~3x longer than JPEG")
-    print("   • Better quality justifies extra time")
-    print("   • Batch processing makes it efficient")
-    print("   • Run overnight for large batches")
-    print()
-
-    print("💡 COST-BENEFIT ANALYSIS:")
-    print("   • Extra storage cost: ~$20/TB")
-    print("   • Extra processing time: 2-3x longer")
-    print("   • Quality improvement: 5-10x better")
-    print("   • Professional results: Priceless")
-    print()
-
-    print("🎯 BOTTOM LINE:")
-    print("   The quality improvement FAR outweighs the costs")
-    print("   Your clients/audience will notice the difference")
-    print("   Professional photographers use RAW exclusively")
-
-
-if __name__ == "__main__":
-    demonstrate_your_pipeline_advantages()
-    show_preset_differences()
-    practical_benefits_for_your_work()
-    your_workflow_recommendation()
-    storage_and_performance_reality()
-
-    print("\n" + "🏆" * 20)
-    print("FINAL ANSWER FOR YOUR QUESTION:")
-    print("🏆" * 20)
-    print()
-    print("📷 SPORTS: Use NEF files")
-    print("👤 PORTRAITS: Use NEF files")
-    print("🦅 WILDLIFE: Use NEF files")
-    print("🏔️  LANDSCAPES: Use NEF files")
-    print("📱 EVERYTHING: Use NEF files when possible")
-    print()
-    print("Your pipeline automatically optimizes everything!")
-    print("Just feed it NEF files and get amazing results!")
-    print("🏆" * 20)
