--- tests/image_processing/__init__.py
@@ -1,1 +1,0 @@
-# Image Processing Tests
