--- tests/preset_analysis/__init__.py
@@ -1,1 +1,0 @@
-# Preset Analysis Tests
