+++ tests/presets/test_format_optimizer.py
