+++ tests/raw/test_rawpy_compatibility.py
