--- tests/raw_processing/__init__.py
@@ -1,1 +1,0 @@
-# RAW Processing Tests
